Fix border line intersections.
Before the right and left colored borders of each cell would intersect
with a 45degree diagonal angle when the colors were different.

To make sure the instersection line is perpendicular (so it doesn't
disrupt the grid's grid-y appearance) use a real border on the right side
and a box-shadow border on the bottom. The intersection is then
perpendicular.