Use CSS rather than react re-renders for smoother animation.
Previous approach: Set scrollRatio state and pass that down through the
header to the logo. This lead to re-rendering the logo, and relying on
CSS transitions to animate between. This was choppy because, even with
throttling, there are so many scroll events and rerenders.

New approach: Rather than react rerendering + CSS transitions, rely
completely on CSS and cut out the rerendering. Dynamically set a CSS
logoScale CSS property on the Logo element (requires some direct DOM
manipulation, and passing down a ref from the Layout component down to
the Logo component). Its a little more complicated, but the result is
a MUCH smoother animation.