WIP. Event delegation to document.body.
Since we want to animate multiple components based off scroll ratio,
AND don't want to pass specific component refs from Layout down to
various components (Logo and AboutBlurb for now), we need a better
system. This commit is a messy proof of concept:

Handle scroll events in the Layout, and fire a custom scrollRatio
event on document.body. Then in the components you want to animate, eg
AboutBlurb, attach a listener to set a CSS property on it's DOM ref
to power its own animation.