Sticky header with no AboutBlurb animation (#40)
* Upgrade gatsby-cli.

* First attempt: half works.

Header is sticky, but footer behaves improperly on the about page
(where the content is tall enough to run offscreen).

* Sticky header, non-sticky footer.

Where the non-sticky footer is at the very bottom of the page even
if the content hight is not the full viewport.

* Calculate content scroll ratio state and pass to Logo.

* Piecewise logo scaling function.

Different behavior for mobile and desktop.

* Use CSS rather than react re-renders for smoother animation.

Previous approach: Set scrollRatio state and pass that down through the
header to the logo. This lead to re-rendering the logo, and relying on
CSS transitions to animate between. This was choppy because, even with
throttling, there are so many scroll events and rerenders.

New approach: Rather than react rerendering + CSS transitions, rely
completely on CSS and cut out the rerendering. Dynamically set a CSS
logoScale CSS property on the Logo element (requires some direct DOM
manipulation, and passing down a ref from the Layout component down to
the Logo component). Its a little more complicated, but the result is
a MUCH smoother animation.

* Add missing propTypes.

* Fix logo size/padding on initial render.

* Add comment with ideas for about blurb scaling.

* Remove TODO comment: We tried and didn't like.

See PR #39 for the exploration.

* Small Prettier style changes.

* Fix initial logo padding.

With css var default value.